# Reduce memory allocations in HttpFileSystembasedFindPackageByIdResource