--- a/src/NuGet.Core/NuGet.Protocol/RemoteRepositories/HttpFileSystemBasedFindPackageByIdResource.cs
+++ b/src/NuGet.Core/NuGet.Protocol/RemoteRepositories/HttpFileSystemBasedFindPackageByIdResource.cs
@@ -557,6 +557,7 @@ namespace NuGet.Protocol
                 builder.Add(parsedVersion);
             }
 
+            builder.Sort();
             return builder.ToImmutable();
         }