<!-- DO NOT MODIFY OR DELETE THIS TEMPLATE. IT IS USED IN AUTOMATION. -->

# Bug

<!-- If this is an engineering change or test change only, you do not need an issue. -->
<!-- Find or create an issue in NuGet/Home and paste the full url. -->
<!-- At the maintainers discretion, multiple changes may apply to a single issue, but only when the PRs are all created within a short period of time. -->
Fixes: https://github.com/NuGet/Home/issues/14095

## Description

Following @ToddGrun's PR: https://github.com/NuGet/NuGet.Client/pull/6265, in particular this comment thread: https://github.com/NuGet/NuGet.Client/pull/6265#discussion_r1980312868

I noticed that the code is creating one `PackageInfo` and `PackageIdentity` and `ContentUri` for every package version the server knows about. However, in a restore, typically only a single package version is needed. So, by creating those `PackageIdentity` and url strings only when needed, we can remove a whole lot of additional allocations over what the other PR did.

In addition to removing `PackageInfo` and associated types, I moved the JSON parsing to System.Text.Json instead of Newtonsoft.Json (which used JObject, no less).  So, that provides us with just most of Todd's PR memory improvements as well.

## PR Checklist

- [x] Meaningful title, helpful description and a linked NuGet/Home issue
- [x] ~Added tests~  code was refactored, no behavior changes
- [x] ~Link to an issue or pull request to update docs if this PR changes settings, environment variables, new feature, etc.~ N/A