@ToddGrun 2025-03-05
I like this PR, it was a good find that the PackageInfo didn't really need to be in this cache.