@ToddGrun 2025-03-06
I'd love to hear about what measurements you ended up taking and how big of an improvement this is