@zivkan 2025-03-06
> I'd love to hear about what measurements you ended up taking and how big of an improvement this is

From: https://github.com/NuGet/NuGet.Client/pull/6265#discussion_r1980768945

|branch|GC stats total allocs|GC Heap Alloc Ignore Free Stacks ConsumeFlatContainer|
|--|--|--|
|dev|252 MB|5,512,529|
|this PR|246 MB|2,139,213|
|my PR|244 MB|669,655|

But keep in mind in the comment, from the other PR, I had the caveat that I amit I pretty much don't know what I'm doing.

Here's a PerfView diff of the dev branch and this PR branch heap alloc view:
![image](https://github.com/user-attachments/assets/a3b6971b-4c0e-4d71-ba86-df4d4fde6336)

Finally, I thought I'd try to use BenchmarkDotNet's MemoryAnalyzer. Here's a patch to the NuGet.Client repo with my benchmark code: 
[benchmark.zip](https://github.com/user-attachments/files/19116333/benchmark.zip)

However, when I run it, I get these results:

| Method   | Mean     | Error     | StdDev    | Ratio | Gen0     | Gen1    | Allocated | Alloc Ratio |
|--------- |---------:|----------:|----------:|------:|---------:|--------:|----------:|------------:|
| Change   | 2.739 ms | 0.0161 ms | 0.0126 ms |  1.00 | 328.1250 | 23.4375 |   1.99 MB |        1.00 |
| Original | 2.656 ms | 0.0246 ms | 0.0205 ms |  0.97 | 316.4063 | 23.4375 |   1.91 MB |        0.96 |

So, this benchmark results make it look like my change has more allocations and is slower than the original, which frankly doesn't make any sense to me whatsoever.

What started as reviewing another PR turned into trying to implement further improvements, and while code changes took probably less than an hour, I've now been working probably 2 full days on collecting perf measurements.  I was behind my assigned work for this sprint before I started any of this, so I just can't justify any more effort.  If this PR is "acceptable" as-is, or with minor changes, I can do that.  But I don't have the capacity to investigate any further why these perf measurements are so out of whack.