@zivkan 2025-03-07
While falling asleep last night, I realized that my benchmark was reusing a single `SourceRepository` instance, which means the same `HttpFileSystemBasedFindPackageByIdResource` instance, which means the warmup would warm the cache, and then the benchmark was only testing the cache-hit code path!  It also explains why my change was slower and using more memory, because the old code pre-cached the content URL and PackageIdentity for every package version, whereas my change allocates those types on demand.

Fixing the benchmark, I get different results:

| Method   | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0      | Gen1     | Allocated | Alloc Ratio |
|--------- |---------:|--------:|--------:|------:|--------:|----------:|---------:|----------:|------------:|
| Change   | 206.4 ms | 4.11 ms | 5.20 ms |  1.00 |    0.03 | 1000.0000 |        - |  20.67 MB |        1.00 |
| Original | 201.5 ms | 4.02 ms | 7.25 ms |  0.98 |    0.04 | 1333.3333 | 666.6667 |  24.77 MB |        1.20 |

I'm still very surprised the changed code is slower, wall clock time. But at least we're now correctly measuring the memory allocation improvements.