@zivkan 2025-03-18
@NuGet/nuget-client tests have been fixed. This is ready to merge now.